requirements.txt
python-telegram-bot==20.7
PyYAML>=6.0
loguru~=0.7.2
psutil>=5.9.0